Lol, ACTUALLY finishing day 9
Got to add some more functionality to my Grid2D class, which is always a good day!

Would very much like to, at some point, structure the problem solver in such a way that it can pass data cleanly from Part 1 to Part 2. That way I wouldn't have to, in this example, re-locate all the lowest points in part 2, because I'd already done that work in part 1.